[DPWBS-1011] Add clang support for TriCore Register Calling Convention
This patch introduces the TriCoreABIInfo and TriCoreTargetCodeGenInfo classes.
The TriCoreABIInfo class is responsible for implementing the target specific
function argument and return value marshalling.

The TriCoreCodeGenInfo class so far only initializes the TriCoreABIInfo instance.

The added test tricore-abi.c demonstrates that all arguments and return values are
lowered as expected.