[DPWBS-1092] fix(InstrInfo): fix ValueType for signed and unsigned immediates, off16
To be able to emit instructions like MOV_dc 0 directly from TableGen
we need to tell TableGen that certain value types are equivalent to our
defined immediates (e.g. i8 should be equivalent to signed/unsigned
imm8.

This commit also adds GlobalISel-compatible ImmLeafs which check that
the immediate value is in range.