[DPWBS-1118] feat(InstructionSelector): constrain destination register for COPY
COPYs are not constrained implicitly but require explicit constraining.
Constrain the destination register to a register class fitting the
selected register bank and the type. The source register is not
constrained as it is constrained once it is encountered at the next use
or definition.

add test to check COPY constraining