[DPWBS-1139] feat(InstructionSelector): select G_FRAME_INDEX to LEA_aac
In the instruction selector G_FRAME_INDEX needs to be
lowered to a target instruction. However, the actual
lowering to a stack-address computation happens later.

Therefore this instruction selection prepares the
actual lowering later. For this purpose we emit
a LEA_aac with offset 0. This instruction will be
changed into the format we need at a later stage.

add selection tests for G_FRAME_INDEX