[DPWBS-1149] feat(RegisterInfo): implement eliminateFrameIndex
Replace frame index operand with $a14 and immediate if offset does not
exceed 16-bits.

Otherwise, materialize the offset into a register and use an explicit
ADDA_aaa $a14, %offset instruction. Replace the frame index with the
result of this instruction and set the offset immediate to 0.

add tests for eliminateFrameIndex

add XFAIL test until we implement emitProlog/emitEpilogue