[DPWBS-1149] feat(FrameLowering): implement emitPrologue
Emit prologue if function uses a stack. Prologue sets up the stack and
frame pointer. Frame pointer initialization is only done when function
uses a frame pointer.

CFI directives for CFA calculation are emitted using the FP if available
or the SP + offset otherwise.

Stack realignment and funclets are not supported yet. An error is
reported if a function requires stack alignment.

There is no need to implement emitEpilogue since SP and FP are
implicitly restored on a return.

Call frame pseudo instructions are eliminated and update the stack
pointer if a call frame has not been reserved. Call frame reservation
is done if no variable-sized objects are on the stack.

When updating the stack pointer and a frame pointer is not available the
CFA offset is also updated.

Emit prologue always sets the FP $a14 to the current SP. A later
optimization should be implemented to eliminate this if $a14 is never
used.

The maximum allowed stack size is 2 ** 32 bytes.

update elim-frame-index tests

add prologue-epilogue tests

add XFAIL test for stack realignment

add XFAIL test for function with no frame pointer (-fomit-frame-pointer)

add invalid test for a too-large stack

add test that CFI directives are printed in assembly

Add XFAIL tests from the GCC bug HDP-1144.
These tests check that instructions using the stack are not scheduled before
prologue instructions. At the moment we are not able to translate these tests
to assembly yet, therefore these tests are added as XFAILs and will be
updated once we can compile them