[DPWBS-1153] feat(TriCoreTargetStreamer) set ELF header E_FLAGS
This change introduces two new classes:

TriCoreTargetStreamer: it provides an abstract interface to the
target specific assembly and object streamers. We need it as a
superclass for our TriCore specific ELF streamer.
For now, we don't need a TriCore specific assembly streamer.

TriCoreTargetELFStreamer: it wraps the actual MCELFStreamer class
and sets the TriCore specfic ELF E_FLAGS.