[DPWBS-1041] fix(TriCoreCallLowering): emit ADJCALLSTACKDOWN/UP pairs around calls.
During the call lowering, we have to determine the amount of stack required to pass the
arguments to a function. It is recorded and communicated to the prolog/epilog insertion
pass via pairs of ADJCALLSTACKDOWN/UP pseudo instructions. Once the stack layout is fixed,
the pseudo instructions are eliminated through the TargetFrameLowering function
eliminateCallFramePseudoInstr, which will be implemented in another patch.

The call-translator.ll test is updated to include the new pseudo instructions around the
calls with the correct stack sizes.