[DPWBS-1041] Add support for floating point types in calling convention
So far we did not have handling for f32/f64 types in the calling convention
description. During a workshop this example was used to explain how to
extend the calling convention and how the call lowering for GlobalISel works.

The tests show that we now can handle floating point arguments and return values
correctly.