[DPWBS-1164] feat(InstructionSelector): select extensions to target instructions
32-bit sign extensions use the extract instruction to extract and
sign-extend the source bits.

64-bit sign extensions can be expressed by a simple mul-by-1 if the
source size is 32-bit. If the source size is smaller than 32-bits
the extract instruction is again used, combined with an arithmetic
right-shift to set the upper 32-bits. If the source size is bigger than
32-bits a simple 32-bit sign-extension on the upper 32-bits is done.

Zero-extension uses the insert instruction to insert zero-bits in the
upper bits of the source. A 64-bit zero-extension moves 0 to the upper
32-bits if the source size is smaller than 32-bits.

G_ANYEXT leaves the upper bits unspecified. This is therefore translated
to a simple COPY instruction if the extension stays in the same register
or an IMPLICIT_DEF with an INSERT_SUBREG instruction otherwise.

add selection tests for extensions