[DPWBS-1164] test(Legalizer): prepare test with generator
Updating the legality rules for legalizer artifacts have a big impact
on many different tests. To update these tests a update script is
usually used. If the tests where hand-written however, this causes very
large non-functional changes as the generator has a different naming
scheme than a human would use.

In order to prepare for this, this commit updates all tests that will be
affected by the legality changes. This way, the commit with the
functional changes only contains the functional changes in the tests as
well.