[DPWBS-1120] feat(SystemOperands): add support for named CSFRs
This commit adds support for using named system registers (e.g. $psw) in
instructions which support a uimm16 immediate operand. Named system
registers are parsed and printed as an immediate. MFCR, MTCR, MFDCR and
MTDCR print the system register names.

For these instructions an immediate input is interpreted and printed as
a named system register, if a register matching the encoding is known.

add assembler tests for named system register operands