[DPWBS-1369] Move load/store handling into C++
We have several SuperTest failures where loads and stores with memory sizes < 32 bit cannot be selected.
The reason is that our TableGen patterns hard-code the source of the smaller type to be a G_TRUNC, which is
not always the case.

Because of limitations in TableGen the general case of smaller types cannot be expressed properly. As a result,
we move the selection into custom C++ code. The existing tests make sure we didn't break anything.