[DPWBS-1369] Fold G_PTR_ADD offset into load/store instruction
If the pointer operand of the load/store is a G_PTR_ADD, and the offset in
G_PTR_ADD is fitting into the immediate operand of the load/store instruction,
we can fold the G_PTR_ADD into the load/store instruction.

Tests are added for every load/store instruction currently selected. Min/max as well as
boundary tests make sure the immediate operand sizes of the load/store instructions are described
correctly in TriCoreInstrInfo.cpp.