[DPWBS-1395] First stab at optimizing constant selection.
We can use shift instructions to move the immediate into place if the pre-shifted immediate fits
into the operand of a single instruction. We prefer this over using the default logic because it
can result in a 16-bit shift instruction. Also, shift does not clobber the PSW register which
allows for better scheduling opportunities in the future.