[DPWBS-1411] feat(PostLegalizerCombiner): add basic post legalizer combiner
Add a basic post legalizer combiner to recombine some of the legalized instructions.
This pass will later be expanded to combine typical generic constructs to certain
target instructions.

For now, extending loads combines are disabled because they require
legalization, while the framework cannot interact with the legalizer
yet.

Indexing loads/stores combines need a target hook to be implemented
before the combiner starts emitting them. However there is currently a
bug in the combiner for indexing loads, causing an llvm_unreachable to
be executed. Until this is fixed, the target hook required will not be
implemented.