[DPWBS-1426] Add support for 'd' inline assembler constraint.
The 'd' inline assembler operand constraint can be used to require
the compiler to use a data register for the given operand.

A 32-bit operand is passed in a 32-bit data register, while a 64-bit operand
is passed via a 64-bit extended data register.

We have to teach clang to recognize the 'd' constraint.

Tests are added for the C -> LLVM IR and LLVM IR -> GMIR translation.