[DBPWS-1395] fix(ExpandPseudoInsts): fix immediate printing by using appropriate data type
This is a follow up on the previous commit which fixed the immediate printing
during frame index elimination. It fixes the immediate printing for constant
materialization.

Use int16_t (and int32_t when appropriate) instead of uint64_t to force a
sign-extension when using .addImm.
This fixes the immediate printing where negative numbers would be printed as
positive numbers, causing our assembly parser to throw an error on our own
assembly.

The cause of this problem was that we zeroed out the upper bits to only get
the relevant bits, but by using uint64_t there was no sign-extension happening
when adding the immediate to the instruction, causing it to be interpreted
as the positive number.