[DPWBS-1385] refactor(Legalizer): make s8, s16 G_PHIs illegal
s8 is never consumed by any of our instructions so make it illegal in
G_PHI.

Although consumed and produced, s1 is not made legal since that is only
used as a carry-bit, which cannot be selected if it where used in G_PHI.

Similarly for s16, although used for half floats, we have no way to
differentiate between half-float and short, so s16 is also considered
illegal.

Once there are float LLTs we can differentiate them and will allow
half-floats for PHIs again.