[DPWBS-1401] Truncate the len argument to memory intrinsics to 32-bits.
LLVM and clang use hard-coded types when creating calls to memory intrinsics in cirtain situations.
In some cases, the type is hard-coded to i64, which is passed in an extended register on TriCore.
When lowering the intrinsic to the corresponding C library function, this results in an ABI mismatch which
breaks at runtime.

For now, we truncate the len argument to 32-bits to work around this problem.